Extract logIterables to utils package
