+++ utils/index.ts
@@ -1,2 +1,3 @@
 export { ansiColors as colors } from "./src/ansiColors";
 export { listAllPkgJsons } from "./src/listAllPkgJsons";
+export { logIterable } from "./src/logIterable";
